second new update to the Remedy script
there are now two function related to "fetched array listings" related to both the directory and file sectors(see line 46-92 for directory sector and line 93-165 for file sector) in which the script uses both of those to generate array strings of either of them for later usage. now the script should be less slower from there.

i've also removed some of the comments 'cuz why not?