sally_bf.bms
the script was alright in its current state, though there were two issues about it.

1. a .bf file contains a signature header that is stored at 4 bytes. however, since a .bf file contains little-endian header structure, i had to "reverse the endianness" by invoking three xmath variables, the first one which is a base for the second one which would be actually used by the script. the third xmath variable might see its usage in the future, though i'm not planning to do anything with that for now so you won't see that in the script.

2. the filexor feature didn't help much of anything and required me to just put on lines that would either disable or enable the feature depending on where putting such lines was necessary for the script to work with a .bf file from some Ubisoft game.

3. the "error" messages (you know, the ones where a script found itself dealing with an unknown .bf file from some Ubisoft game) were confusing at best and dismissive at worst. not to mention the "version handling" part itself was very strict as i worked with whatever .bf files i had at the time and so there wasn't much reason for me to pay attention to these kinds of things. it only became an issue when i spotted up an conversation about it at some hcs64 thread, with someone in particular expressing zir* confusion over such error messages.

with this in mind, i started work on this script just yesterday with just enough .bf files that i've grabbed myself to help me out with the undertaking of addressing the script's most obvious issues as outlined above. but these issues weren't going to be solved by just a few lines, the script would also be fleshed out to accommodate many .bf files that i could find out of any Ubisoft game whatsoever.

with that being said, the focus on this script will be on the games that Ubisoft has developed and published for release on GameCube and Wii platforms. although some of these Ubisoft games aren't "exclusives" to either of these platforms by any sense of the word, nevertheless i thought it would be nice to cover them up so the script could have "almost everything" covered regarding the BIG/BUG format as deployed by Ubisoft in some of their games.

another new feature that wasn't there until now is that it can now get through as many "file indices" as the .bf file has. having dozens of "file indices" to seek through without issue is a feature of the BIG/BUG format that i haven't seen documented anywhere else, until i've somehow managed to find a game (the Wii version of My Word Coach) that uses such a feature through one of its files (spree.bf).

* - i'm pretty sure my usage of these pronouns are correct.