sally_bf.bms
fixed an potentially glaring oversight that would have massive consequences when parsing any "file index" that the .bf file uses. that is to say, decrypt "file index" headers at all times instead of when the "BUG" signature is on.

one thing i forgot to mention though is that most of the BIG/BUG(.bf) format is now handled by taking advantage of quickbms' MEMORY_FILE feature. this was one of the very first things i've sought to implement to address the issue of having the script abuse "filexor" feature to such an extreme, not to mention having the script to go back and forth between many "offsets" all at once to handle tons of things regarding said format's structure is not really something i'd call "practical" in every sense of the word. as such, if the .bf file comes starts with "BIG" signature just about every aspect of said format is covered into about 5 MEMORY_FILEs each.

for decryption stuff (which is reserved for .bf files with "BUG" signature) a variable covering one of the format headers is used as a key, and thus many memory allocations are made in place to ensure the decryption process goes off without a hitch.

both of these methods were implemented to ensure the format structure is read in a straight-forward manner, which explains the presence of many "file_number" words you get to see around the entire script.

as for the Xbox versions of Beyond Good and Evil (one .bf file split into several parts with .bf# extension, # being the split number that starts with 0 alongside some file size funk in these splitted files) and Peter Jackson's King Kong: The Official Game of the Movie (same case as BG&E except the extension is now .###, ### starting with 1), i can assure you that these will be implemented whenever possible.