sally_bf.bms
big changes here, not only relating to the "BIG"/"BUG" archives in question but also extending support to other .bf files whose signatures are "ABE" and "BF64". however, only the "BIG"/"BUG" archives are able to be extracted with this script as i am not finished with those two yet.

to cap things off, "version" is now "bigfile_parameters" because these are the ones that impact the "file archival" structure of a .bf file. the Xbox 360 version of King Kong can now be supported with at least three files (KKTextures.bf, KKMaps.bf, Sound_Common.bf) benefiting from this script. and finally i managed to nail down the "ABE" file structure. work still needs to be done with the "BF64" file structure but i need to find the will for this.