squaresoft-square.enix_sz_dat.bms
FFX (PS2) movie files (all stored on the latter half of the game disc) can now be "fully" handled... so long as the sound parts for these files exist. all PS2 releases of FFX has at least one movie file whose "sound" chunks emply a wide hybrid of "compressed" PCM and "untouched" PCM.

however, i found out (through reverse-engineering the whole thing on GHIDRA 9.1.2 plus some add-on i got from somewhere aroudn github that covers the PS2 CPU assembly language) that these "compressed" PCM chunks employed something rather simple: there is a "central byte" that poses as one part of what i'm going to call a "sample number". said byte contains a control symbol (1-bit) that is carved to a much higher place that dictates whether or not a "secondary byte" is needed. if it is, then the "central byte" and "secondary byte" are combined to make "sample number" into an entire entity. however, "central byte" takes the utmost priority in being read, and it's being read through a single variable that determines the position (a "variable" can be any number) of the "central byte" to be read. i call this the "central byte position".

in the midst of all this, massive calculations are made, all of which are handled by another single variable. this variable is the core of a "sample number" and needs either one ("central byte") or two bytes ("central byte" and "secondary byte") to make extensive calculations so "sample number" can look more handsome. for that end, "sample number" takes a "central byte" and expands it by left-shifting it to 25 and then waits if the "control bit" can say if there's a "secondary byte" for "sample number" to handle. if there isn't, "sample number" right-shifts itself to 25 and "central byte position" walks one step. otherwise, "central byte position" takes two steps, a "secondary byte" is obtained, "sample number right-shifts itself to 17 and appends the value of that "secondary byte" into itself.

when all that is done, a "sample number" has its number added by an "assisting number" (which is set to 0 upon startup of "compressed" PCM decoding that i'm explaining here). from there on, "sample number" left-shifts itself by 16, right-shifts itself by 16 and the "assisting number" copies the value of "sample number. after all this, the game writes the "sample number" into memory and decodes said number into real-time sound playback of an FMV.

with that being said, something like this is rather slow. i'm well aware of any criticism (either incoming or already vocalized in some way) that comes as a result of this script's rather slow performance. but to optimize the script's perfomance would mean to recode some parts of the script into C code encapsulated into MEMORY_FILE binary parts so calldll (it's a command that allows custom plugins to be used into quickbms, said command allows for memory file code compilation like it was C code) could be effectively used (example - calldll MEMORY_FILE20 some_function "tcc" RET some_return_value). and i'm not exactly experienced in C code so don't count on it...

on a lesser note, ".cam" files can now be extracted without this "file already exists" prompt since we're dealing with other chunk types here.