squaresoft-square.enix_sz_dat.bms
added full audio support for Kingdom Hearts (PS2) FMV files.

said audio contains at least two modes: "first mode" which contains 512 bytes of "untouched" PCM16 samples, and "second mode" which employs the same "DPCM" audio algorithm that was used in FFX (PS2) FMV files (all of which use an entirely different format which is also handled here) except it's now entirely reliant on two outside values (frame size and "predicted sample", respectively) so the output sound playback that goes within an FMV file (during your playthrough that is) is controlled and accounted for. this is why i tweaked this algorithm i figured out literally three days ago so it could be used for decoding audio tracks out of a Kingdom Hearts (PS2) FMV file (opn.dat for example) without breaking existing compatibility with FFX (PS2) FMV files.

all of this is handled within the "type 4" chunk of this .dat format Kingdom Hearts (PS2) uses (see line 73). this "type 4" chunk is capable of storing at least an entire "untouched" frame that (depending on the codec number which can be seen at lines 840-848) contains full uncensored audio data. the codec numbers are as follows:

0xffffff00 - 16-bit pulse code modification
0xffffff01 - Dolby Digital AC-3
0xffffff02 - DPCM
0xffffff03 - VSF container, used in Front Mission 5 (will be handled one day)

codec number 0xffffff02 needs to be handled differently because the way it's designed is that DPCM decoding must be done within the chunk instead. for that end, multiple "sub-blocks" exists whose "handling" is dictated by the two modes which are outlined above.

i believe that's pretty much it really. what's left out of this script is the video part, so maybe i won't find myself struggling with it this time around now that the COVID-19 pandemic is around the corner.