sally_bf.bms
this is a massive MASSIVE update.
most LyN-based files can be extracted without issue thanks to a LOT of stuff i had to get through to ensure the script actually worked at all.

but first things first, more comments. they try to give the reader some insight on how certain parts of the script work. it's not much but you can expect a lot of comments in the following updates of this script if there will ever be one! but more on that later.
secondly, the part where the script tries to detect if a "big bug" file is split into several parts each is now in its proper place instead of at the start of the script. wanted to extract some just dance videos from two bigfiles but couldn't? now you can!
and finally, the script has been fine-tuned for extracting all files residing in ABE bigfiles. this is by far the biggest undertaking when it came to updating this script to a new version. proper LyN file extraction is now possible. mostly.

that's it for now. i did what i could within the limits of this programming language but i'm afraid this script isn't going to be enough in most cases. see you later i guess.